==> Archive 04-2020/0_EDM_simplex_vs_block_lnlp.Rmd <==
---
title: "0_EDM_simplex_vs_block_lnlp"
author: "Alison Iles"
date: "9/30/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#load the necessary packages. 
Note that the `echo = FALSE` parameter prevents printing of the R code.

```{r, echo=FALSE}
library(rEDM)
library(reshape2)
library(ggplot2)
```

#set up the data block
```{R}

load("Data/Rdata/block_data.Rdata")  

mpg <- c() 
  for(stk in 1:length(block_data)) { 
      mpg[stk] <- block_data[[stk]]$mpg[1]
  } 
  
u <- 3   # isolate the third mpg
    data <- block_data[c(mpg==u)]
    name <- as.character(data[[1]]$mpg[1])
    
  var <- c("eff_n") #focal variable
    
    #merge the time series of the stocks together
    keys <-c("yr", var)
    merged_data <-setNames(do.call(mapply, c(FUN=c, lapply(data, `[`, keys))), keys)
    merged_data <- as.data.frame(merged_data)
    valid <- is.finite(merged_data[,2])
    block <- merged_data[valid,]  
    
    #List in lib the begin and end points of each stock and the break points within stocks
    lib <- matrix(NA, nrow = length(which(diff(block$yr)!=1))+1, ncol = 2)
    lib[,1] <- c(1, which(diff(block$yr)!=1)+1)
    lib[,2] <- c(which(diff(block$yr)!=1), nrow(block))

```    

Run simplex and block_lnlp with various embedding dimensions. plot rho as a function of E.
```{R}
#Simplex
simplex_output <- simplex(block[,2], lib = lib, pred = lib, E = 1:20, exclusion_radius = 0, silent = TRUE)

#block_lnlp             
lag_data <- block[var] #make lagged data block
lag_block <- make_block(lag_data, t = NULL, max_lag = 20, tau = 1, lib = lib, restrict_to_lib = TRUE)
lnlp_E_rho <- matrix(NA, nrow = 20, ncol = 2) #empty matrix to store results from loop through different E
        for(e in 1:20) { 
           lnlp_output <- block_lnlp(lag_block[2:(e+1)], lib = lib, pred = lib, method="simplex")
           lnlp_E_rho[e,1] <- e
           lnlp_E_rho[e,2] <- lnlp_output$rho
        }         
lnlp_E_rho <- as.data.frame(lnlp_E_rho)
            
#Plot the results (simplex output vs. block_lnlp output):
ggplot() + 
  geom_line(data = lnlp_E_rho, aes(x = lnlp_E_rho[,1], y = lnlp_E_rho[,2]), color = "red", size=2) +
  geom_line(data = simplex_output, aes(x = E, y = rho), color = "blue") +
  geom_line() + theme_bw() +
  xlab('E') +
  ylab('rho')
```
