==> Archive 04-2020/1_EDM_Preprocess_data.Rmd <==
---
title: "1_EDM_Preprocess_data"
author: "Alison Iles"
date: "6/27/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(rEDM)
library(reshape2)
library(ggplot2)
library(gridExtra)
library(xtable)
library(rlist)
library(tidyr)
library(tidyr)
library(pracma) #detrending function
library(zoo) #na.approx function
```

#load Chinook stock data and harvest data
```{R}
rm(list=ls())  
data <- read.csv("Data/csv_data/CORE_2020_data_stock.csv")
```

#normalize

```{R}
 # To focus on one stock to trouble shoot the code
 #stock_df <- subset(data, stk == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers") 
 

# This code normalizes by cycle line. Chinook don't demonstrate a strong cycle line, so it is set to '1', normalize on a 1 year cycle 
    normalize_by_cycle_line <- function(ts)
{
    n <- length(ts)
    means <- rep.int(NA, times = 4) #replicate NA 4 times
    sds <- rep.int(NA, times = 4)
    mu <- rep.int(NA, times = n)
    sigma <- rep.int(NA, times = n)
    for(k in 1:1) #for each cycle line (every 4 years the population cycles in sockeye, I've changed it to 1 because there are no cycle lines for chinook)
    {
        index <- seq(from = k, to = n, by = 1) #changed by=4 to by=1.
        means[k] <- mean(ts[index], na.rm = TRUE) #mean of every 4th element of the time series
        sds[k] <- sd(ts[index], na.rm = TRUE) #standard dev.
        mu[index] <- means[k]
        sigma[index] <- sds[k]
    }
    ts <- (ts - mu) / sigma  #normalize by cycle line
    df <- data.frame(cbind(ts, mu, sigma))
    return(df)
    }
    


     preprocess_stock <- function(stock_df)
    {
        n <- NROW(stock_df)  #n is the number of observations for the stock
         
        stock_df$ret <- stock_df$rec3 + c(NA, stock_df$rec4[1:(n-1)]) + c(NA, NA, stock_df$rec5[1:(n-2)]) #total returns in a given year are calculated by adding 3 year old recruits plus 4 year old recruits from the previous brood year, plus 5 year olds from two years prior...etc. This value is aligned to the rec3 brood year, not the year they are actually returning together. 
       
        # Normalize on a 1 year cycle line (i.e. NO cycle line!)
        temp <- normalize_by_cycle_line(stock_df$spawn_tot)
        stock_df$salm.effn.0 <- temp$ts #normalized number of effective spawners 
        #stock_df$eff_mu <- temp$mu
        #stock_df$eff_sigma <- temp$sigma
        
        temp <- normalize_by_cycle_line(stock_df$rec3)
        stock_df$salm.rec3n.0 <- temp$ts #normalized 3 year old recruits
        #stock_df$rec3_mu <- temp$mu
        #stock_df$rec3_sigma <- temp$sigma
              for(i in c(0:2)) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, stock_df$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = 1))
              df[1] <- stock_df$salm.rec3n.0[index_in_spp_data]
              df.colnames <- paste("salm.rec3n", i, sep = ".")
              stock_df[, df.colnames] <- df
              }
        temp <- normalize_by_cycle_line(stock_df$rec4)
        stock_df$salm.rec4n.0 <- temp$ts #normalized 4 year old recruits
        #stock_df$rec4_mu <- temp$mu
        #stock_df$rec4_sigma <- temp$sigma
              for(i in 0:1) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, stock_df$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = 1))
              df[1] <- stock_df$salm.rec4n.0[index_in_spp_data]
              df.colnames <- paste("salm.rec4n", i, sep = ".")
              stock_df[, df.colnames] <- df
              }

        temp <- normalize_by_cycle_line(stock_df$rec5)
        stock_df$salm.rec5n.0 <- temp$ts #normalized 5 year old recruits
        #stock_df$rec5_mu <- temp$mu
        #stock_df$rec5_sigma <- temp$sigma
              for(i in -1:0) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, stock_df$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = 1))
              df[1] <- stock_df$salm.rec5n.0[index_in_spp_data]
              df.colnames <- paste("salm.rec5n", i, sep = ".")
              stock_df[, df.colnames] <- df
              }
        
        temp <- normalize_by_cycle_line(stock_df$recspn) #normalized total recruits per spawner
            stock_df$salm.recspnn.0 <- temp$ts 
        temp <- normalize_by_cycle_line(stock_df$recspn3) #normalized 3 year old recruits per spawner
            stock_df$salm.recspn3n.0 <- temp$ts  
        temp <- normalize_by_cycle_line(stock_df$recspn4) #normalized 4 year old recruits per spawner
            stock_df$salm.recspn4n.0 <- temp$ts 
        temp <- normalize_by_cycle_line(stock_df$recspn5) #normalized 5 year old recruits per spawner
            stock_df$salm.recspn5n.0 <- temp$ts 
        
        temp <- normalize_by_cycle_line(stock_df$rec)
        stock_df$salm.recn.0 <- temp$ts #normalized total recruits
        #stock_df$rec_mu <- temp$mu
        #stock_df$rec_sigma <- temp$sigma
        
        temp <- normalize_by_cycle_line(stock_df$ret)
        stock_df$salm.retn.0 <- temp$ts #normalized returns
        #stock_df$ret_mu <- temp$mu
        #stock_df$ret_sigma <- temp$sigma
        
        temp <- normalize_by_cycle_line(stock_df$harvest)
        stock_df$harv.SPSS.0 <- temp$ts #normalized harvest with name in the same format as ESU level harvest data
        #stock_df$harv_mu <- temp$mu
        #stock_df$harv_sigma <- temp$sigma
        
        return(stock_df)
    }   
     
     
     
stock_data <- split(data, data$stk)    
stock_data <- lapply(stock_data, preprocess_stock) 



#stock_df <- subset(stock_data[["Camas Creek"]]) 
#Stock data with interpolated missing values
interpolate_stock <- function(stock_df){
        Dn <- stock_df$stk[1]
        D <- stock_df$salm.rec4n.0 #normalized 4 year old recruits per spawner
              #remove early years for as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ){D[1:5] <- NA}

              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 58
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek"| Dn == "Sulphur Creek" | Dn == "Valley Creek" | Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:66] <- na.approx(D[1:66])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:66] <- na.approx(D[1:66])}
              
        stock_df$salm.rec4n.0 <- D      
        
        D <- stock_df$salm.rec5n.0 #normalized 5 year old recruits per spawner
              #remove early years for Wenaha as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ) {D[1:5] <- NA}
              
              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 57
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek" | Dn == "Sulphur Creek" | Dn == "Valley Creek"| Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:65] <- na.approx(D[1:65])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:65] <- na.approx(D[1:65])}
              
        stock_df$salm.rec5n.0 <- D 
        
        D <- stock_df$salm.recspn4n.0 #normalized 4 year old recruits per spawner
              #remove early years for as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ){D[1:5] <- NA}

              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 58
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek"| Dn == "Sulphur Creek" | Dn == "Valley Creek" | Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:66] <- na.approx(D[1:66])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:66] <- na.approx(D[1:66])}
              
        stock_df$salm.recspn4n.0 <- D      
        
        D <- stock_df$salm.recspn5n.0 #normalized 5 year old recruits per spawner
              #remove early years for Wenaha as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ) {D[1:5] <- NA}
              
              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 57
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek" | Dn == "Sulphur Creek" | Dn == "Valley Creek"| Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:65] <- na.approx(D[1:65])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:65] <- na.approx(D[1:65])}
              
        stock_df$salm.recspn5n.0 <- D 
        return(stock_df)
    }   

interp_stock_data <- lapply(stock_data, interpolate_stock) 
```

#load large scale (ESU scale) environmental and species data and mid scale (MPG scale) river flow data

```{R}
env_data <- read.csv("Data/csv_data/SRSS_env_seasonal.csv") #for monthly abiotic variables use 'SRSS_env_monthly.csv'
esu_data <- read.csv("Data/csv_data/CORE_2020_data_ESU.csv") #read.csv("Data/csv_data/CORE_2020_data_ESU_NAforZeros.csv")
mpg_data <- read.csv("Data/csv_data/CORE_2020_data_MPG.csv")

# Change the peak flow date to day of year
startyrdate <- as.Date(paste(as.character(mpg_data$year),"-01-01",sep=""))
peakflowdate <- as.Date(mpg_data$flow.peakdate)
mpg_data$flow.peakday <- as.numeric(peakflowdate - startyrdate)
mpg_data$flow.peakdate <- NULL

# use this to focus on one stock to trouble-shoot the function code:
#stock_df <- subset(data, stk == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers") 


normalize <- function(block)
{
    if(NCOL(block) > 1)
    {
        n <- NROW(block)
        means <- sapply(block, mean, na.rm = TRUE)
        sds <- sapply(block, sd, na.rm = TRUE)
        return((block - matrix(rep(means, each = n), nrow = n)) / 
                   matrix(rep(sds, each = n), nrow = n))
    }
    else
        return((block - mean(block, na.rm = TRUE)) / sd(block, na.rm = TRUE))
}


make_block <- function(stock_df, env_data, esu_data)
    {
        env_names <- names(env_data)[2:length(names(env_data))]  
        env_data[, 2:NCOL(env_data)] <- normalize(env_data[2:NCOL(env_data)]) 
        
        csl_names <- c("csl.haulout", "csl.males", "csl.males6", "csl.bonn", "csl.COL", "csl.WA", "csl.SalishSea") 
        csl_data <- normalize(esu_data[, csl_names])
        
        CSL_pups_data <- normalize(esu_data[, "csl.pups"])
        
        ssl_names <- c("ssl.WA", "ssl.SalishSea", "ssl.WCVIBC", "ssl.SEAK", "ssl.WAK", "ssl.rookORraw", "ssl.bonn")          
        ssl_data <- normalize(esu_data[, ssl_names])
        
        hseal_names <- c("hseal.COL", "hseal.WA", "hseal.SalishSea", "hseal.WCVIBC")
        hseal_data <- normalize(esu_data[, hseal_names])
    
        orca_names <- c( "orca.SRKWpodJ", "orca.SRKWpodK", "orca.SRKWpodL", "orca.SRKWpodKL", "orca.SRKWpodJKL", "orca.SRKWdeathsJKL", "orca.SRKWbirthsJKL", "orca.NRKW", "orca.SEAK", "orca.WAK")
        orca_data <- normalize(esu_data[, orca_names]) 
        
        harv_names <- c("harv.PACtot", "harv.CRsport", "harv.COL", "harv.ORWAsport", "harv.ORWA", "harv.SEAK", "harv.NBC", "harv.WCVI")
        harv_data <- normalize(esu_data[, harv_names]) 
        
        hatch_names <- c("hatch.spring", "hatch.summer", "hatch.total", "hatch.SNAK", "hatch.COL", "hatch.CAOR", "hatch.WA", "hatch.SAL", "hatch.BC", "hatch.SEAK", "hatch.WAK", "hatch.sum", "hatch.all")
        hatch_data <- normalize(esu_data[, hatch_names]) 
        
      
 
        # line up species data at ESU level
      
        # offset hatchery data from 1 to 3 years after brood year - assuming the hatchery releases age 2 smolts and they compete with the naturally spawned smolts starting at age 2
        for(i in c(1:3)) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, esu_data$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(hatch_data)))
              df[index_in_stock_df,] <- hatch_data[index_in_spp_data, ]
              df.colnames <- paste(hatch_names, i, sep = ".")
              stock_df[, df.colnames] <- df
              
        }
        
        # offset harvest data from 4 to 5 years after brood year 
        for(i in c(4:5)) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, esu_data$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(harv_data)))
              df[index_in_stock_df,] <- harv_data[index_in_spp_data, ]
              df.colnames <- paste(harv_names, i, sep = ".")
              stock_df[, df.colnames] <- df
              
        }
        
        # offset pinniped data from 2 to 5 years after brood year - assuming seals and sea lions eat smolts and returns
        for(i in c(2:5)) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, esu_data$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(csl_data)))
              df[index_in_stock_df,] <- csl_data[index_in_spp_data, ]
              df.colnames <- paste(csl_names, i, sep = ".")
              stock_df[, df.colnames] <- df
        }
            for(i in c(2:5)) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, esu_data$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(ssl_data)))
              df[index_in_stock_df,] <- ssl_data[index_in_spp_data, ]
              df.colnames <- paste(ssl_names, i, sep = ".")
              stock_df[, df.colnames] <- df
            }
            for(i in c(2:5)) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, esu_data$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(hseal_data)))
              df[index_in_stock_df,] <- hseal_data[index_in_spp_data, ]
              df.colnames <- paste(hseal_names, i, sep = ".")
              stock_df[, df.colnames] <- df
            }

  
        # offset CSL pup data from 2 to 5 years after brood year. Add 6 years as only 6+ year old sea lion males swim up to Oregon and eat adult salmon (ref in Chasco paper)
        for(i in c(8:11)) { 
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, esu_data$year)
              index_in_stock_df <- 1:length(desired_years)
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(CSL_pups_data)))
              df[index_in_stock_df,] <- CSL_pups_data[index_in_spp_data ]
              df.colnames <- paste("csl.pups", (i-6), sep = ".")
              stock_df[, df.colnames] <- df
        }
        
        # line up orca data
        for(i in 4:5) { # offset orca data from 4 to 5 years after brood year - assuming orcas just eat the returning fish
              desired_years <- stock_df$year + i
              index_in_spp_data <- match(desired_years, esu_data$year)
              index_in_stock_df <- 1:length(desired_years)
                     
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(orca_data)))
              df[index_in_stock_df,] <- orca_data[index_in_spp_data, ]
              df.colnames <- paste(orca_names, i, sep = ".")
              stock_df[, df.colnames] <- df
        }

         
        # line up ocean environmental data at ESU level
        # offset environmental data from 2 to 5 years after brood year as these are the ocean years
        for(i in 2:5) { 
              desired_years <- stock_df$year + i
              index_in_env_data <- match(desired_years, env_data$year)
              index_in_stock_df <- 1:length(desired_years)
        
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(env_data)-1))
              df[index_in_stock_df,] <- env_data[index_in_env_data, 2:NCOL(env_data)]
              df.colnames <- paste(env_names, i, sep = ".")
              stock_df[, df.colnames] <- df
        }
        

        
        # line up river flow data at MPG level
        desired_mpg <- as.character(stock_df$mpg[[1]])
        mpg_tempdata <- mpg_data[mpg_data$mpg==desired_mpg,]
        mpg_tempdata$mpg <- NULL
        mpgdata_names <- c("flow.mean", "flow.peak", "flow.gageht", "flow.peakday")
        mpg_tempdata[, 2:NCOL(mpg_tempdata)] <- normalize(mpg_tempdata[, mpgdata_names]) 
        
        for (i in 0:5) {
              desired_years <- stock_df$year + i
              index_in_mpg_data <- match(desired_years, mpg_tempdata$year)
              index_in_stock_df <- 1:length(desired_years)
              
              df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(mpg_tempdata)-1))
              df[index_in_stock_df,] <- mpg_tempdata[index_in_mpg_data, 2:NCOL(mpg_tempdata)]
              df.colnames <- paste(mpgdata_names, i, sep = ".")
              stock_df[, df.colnames] <- df
        }
        
        
        return(stock_df)
    }



block_data <- lapply(interp_stock_data, function(stock_df) {make_block(stock_df, env_data, esu_data)})


    
# save and return
    save(block_data, file = "Data/Rdata/block_data.Rdata")

```


# make a data block with no lags
```{R}
rm(list=ls())  
data <- read.csv("Data/csv_data/CORE_2020_data_stock.csv")

library(tidyr)
library(stringr)
library(ggplot2)

normalize <- function(block)
{
    if(NCOL(block) > 1)
    {
        n <- NROW(block)
        means <- sapply(block, mean, na.rm = TRUE)
        sds <- sapply(block, sd, na.rm = TRUE)
        return((block - matrix(rep(means, each = n), nrow = n)) / 
                   matrix(rep(sds, each = n), nrow = n))
    }
    else
        return((block - mean(block, na.rm = TRUE)) / sd(block, na.rm = TRUE))
}

    preprocess_stock <- function(stock_df)
    {
        n <- NROW(stock_df)  #n is the number of observations for the stock
         
        stock_df$ret <- stock_df$rec3 + c(NA, stock_df$rec4[1:(n-1)]) + c(NA, NA, stock_df$rec5[1:(n-2)]) #total returns in a given year are calculated by adding 3 year old recruits plus 4 year old recruits from the previous brood year, plus 5 year olds from two years prior...etc. This value is aligned to the rec3 brood year, not the year they are actually returning together. 
       
        # Normalize on a 1 year cycle line (i.e. NO cycle line!)
        stock_df$salm.effn <- normalize(stock_df$spawn_tot) #normalized number of effective spawners
        stock_df$salm.rec3n <- normalize(stock_df$rec3) #normalized 3 year old recruits
        stock_df$salm.rec4n <- normalize(stock_df$rec4) #normalized 4 year old recruits
        stock_df$salm.rec5n <- normalize(stock_df$rec5) #normalized 5 year old recruits
        stock_df$salm.recn <- normalize(stock_df$rec) #normalized total recruits
        stock_df$harv.SPSS <- normalize(stock_df$harvest) #normalized harvest with name in the same format as ESU level harvest data
        stock_df$salm.recspnn <- normalize(stock_df$recspn) #normalized total recruits per spawner
        stock_df$salm.recspn3n <- normalize(stock_df$recspn3) #normalized 3 year old recruits per spawner
        stock_df$salm.recspn4n <- normalize(stock_df$recspn4) #normalized 4 year old recruits per spawner
        stock_df$salm.recspn5n <- normalize(stock_df$recspn5) #normalized 5 year old recruits per spawner
        stock_df$retn <- normalize(stock_df$ret) #normalized total returns
        
        return(stock_df)
    }   
     
stock_data <- split(data, data$stk)
#stock_df <- subset(data, stk == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers") 

stock_data <- lapply(stock_data, preprocess_stock) 


#stock_df <- subset(stock_data[["Camas Creek"]]) 
#Stock data with interpolated missing values
interpolate_stock <- function(stock_df){
        Dn <- stock_df$stk[1]
        D <- stock_df$salm.rec4n #normalized 4 year old recruits per spawner
              #remove early years for as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ){D[1:5] <- NA}

              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 58
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek"| Dn == "Sulphur Creek" | Dn == "Valley Creek" | Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:66] <- na.approx(D[1:66])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:66] <- na.approx(D[1:66])}
              
        stock_df$salm.rec4n <- D      
        
        D <- stock_df$salm.rec5n #normalized 5 year old recruits per spawner
              #remove early years for Wenaha as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ) {D[1:5] <- NA}
              
              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 57
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek" | Dn == "Sulphur Creek" | Dn == "Valley Creek"| Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:65] <- na.approx(D[1:65])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:65] <- na.approx(D[1:65])}
              
        stock_df$salm.rec5n <- D 
        
        D <- stock_df$salm.recspn4n #normalized 4 year old recruits per spawner
              #remove early years for as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ){D[1:5] <- NA}

              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 58
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek"| Dn == "Sulphur Creek" | Dn == "Valley Creek" | Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:66] <- na.approx(D[1:66])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:66] <- na.approx(D[1:66])}
              
        stock_df$salm.recspn4n <- D      
        
        D <- stock_df$salm.recspn5n #normalized 5 year old recruits per spawner
              #remove early years for Wenaha as there is a large gap soon after the early years.
              if(Dn == "Grande Ronde Upper Mainstem" ){D[1:6] <- NA}
              if(Dn == "Wenaha River" ) {D[1:5] <- NA}
              
              #interpolate missing years for those stocks missing years in middle of their range
              rangeend <- 57
              if(Dn == "Camas Creek" | Dn == "Loon Creek" | Dn == "Marsh Creek" | Dn == "Sulphur Creek" | Dn == "Valley Creek"| Dn == "Yankee Fork") {D[1:rangeend] <- na.approx(D[1:rangeend])}
              if(Dn == "Lostine River") {D[1:65] <- na.approx(D[1:65])}
              if(Dn == "Salmon River Lower Mainstem below Redfish Lake" ) {D[1:62] <- na.approx(D[1:62])}
              if(Dn == "Wallowa River, Hurricane Creek, Bear Creek, and Lostine Rivers" ) {D[1:65] <- na.approx(D[1:65])}
              
        stock_df$salm.recspn5n <- D 
        return(stock_df)
    }   


stock_data <- lapply(stock_data, interpolate_stock)


save(stock_data, file = "Data/Rdata/stock_data_no_lags.Rdata")

env_data <- read.csv("Data/csv_data/SRSS_env_seasonal.csv")
esu_data <- read.csv("Data/csv_data/CORE_2020_data_ESU.csv")
mpg_data <- read.csv("Data/csv_data/CORE_2020_data_MPG.csv")

# Change the peak flow date to day of year
startyrdate <- as.Date(paste(as.character(mpg_data$year),"-01-01",sep=""))
peakflowdate <- as.Date(mpg_data$flow.peakdate)
mpg_data$flow.peakday <- as.numeric(peakflowdate - startyrdate)
mpg_data$flow.peakdate <- NULL

make_block <- function(stock_df, env_data, esu)
    {
        # line up environmental data at ESU level        
        env_names <- names(env_data)[2:length(names(env_data))]  
        desired_years <- stock_df$year
        index_in_env_data <- match(desired_years, env_data$year)
        index_in_stock_df <- 1:length(desired_years)
  
        df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(env_data)-1))
        df[index_in_stock_df,] <- env_data[index_in_env_data, 2:NCOL(env_data)]
        df <- normalize(df)
        df.colnames <- env_names
        stock_df[, df.colnames] <- df
        
        # line up species data at ESU level     
        esu_names <- names(esu_data)[2:length(names(esu_data))]  
        esu_data[, 2:NCOL(esu_data)] <- normalize(esu_data[, 2:NCOL(esu_data)])

        desired_years <- stock_df$year
        index_in_esu_data <- match(desired_years, esu_data$year)
        index_in_stock_df <- 1:length(desired_years)
  
        df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(esu_data)-1))
        df[index_in_stock_df,] <- esu_data[index_in_esu_data, 2:NCOL(esu_data)]
        df.colnames <- esu_names
        stock_df[, df.colnames] <- df
        
        # line up river flow data at MPG level
        desired_mpg <- as.character(stock_df$mpg[[1]])
        mpg_tempdata <- mpg_data[mpg_data$mpg==desired_mpg,]
        mpg_tempdata$mpg <- NULL
        mpgdata_names <- c("flow.mean", "flow.peak", "flow.gageht", "flow.peakday")
        mpg_tempdata[, 2:NCOL(mpg_tempdata)] <- normalize(mpg_tempdata[, mpgdata_names]) 

        desired_years <- stock_df$year
        index_in_mpg_data <- match(desired_years, mpg_tempdata$year)
        index_in_stock_df <- 1:length(desired_years)
        
        df <- data.frame(matrix(NA, nrow = length(desired_years), ncol = NCOL(mpg_tempdata)-1))
        df[index_in_stock_df,] <- mpg_tempdata[index_in_mpg_data, 2:NCOL(mpg_tempdata)]
        df.colnames <- mpgdata_names
        stock_df[, df.colnames] <- df
        
        return(stock_df)
    }



block_data_no_lags <- lapply(stock_data, function(stock_df) {make_block(stock_df, env_data, esu_data)})


    
# save and return
    save(block_data_no_lags, file = "Data/Rdata/block_data_no_lags.Rdata")

```

# plot the data

```{R}

#concatenate the time series of each variable together
    keys <-c(names(stock_data[[1]]))
    block <-setNames(do.call(mapply, c(FUN=c, lapply(stock_data, `[`, keys))), keys)
    block <- as.data.frame(block)
    rownames(block) <- NULL

keycol <- "variable"
valuecol <- "value"
gathercols <- colnames(block[4:26])

D <- gather(block, keycol, valuecol, gathercols)
     colnames(D) <- c("mpg", "stk","year","key","value")

varcat <- unique(D$stk) 
for(c in c(1:length(varcat))) {
     D_plot <- D[D$stk==varcat[c],]     
     p1 <- ggplot(D_plot, aes(x=year, y=value) )+
         geom_line(aes(col=key)) +
         geom_point(aes(col=key)) +
         geom_hline(aes(yintercept=0), colour='#999999') +
         theme_bw() + 
         guides(col = guide_legend(ncol=1)) +
         labs(title= varcat[c], subtitle = "", x="year", y="normalized value") 
         #facet_wrap(vars(key)) 
     print(p1)
}


```
